使用nohup &启动后台进程；通过 > log文件 2>1& 将控制台输出输出到文件；并将pid输出到文件用户kill进程时使用

+++ startOrStopABash/usage.sh
@@ -3,12 +3,17 @@ app="error-bash"
 
 #后台启动bash：将输出打印到日志文件，生成pid文件（关键点 nohup &; 2>&1）
 function start(){
+    
     if [ -f "$app.pid" ];then
         echo "$app areadly start"
     else 
-        echo "start $app"
-        nohup sh error.sh > log.txt 2>&1 &
-        printf '%d' $! > $app.pid
+        if [ -f $1 ];then 
+            echo "start $app"
+            nohup sh $1 > log.txt 2>&1 &
+            printf '%d' $! > $app.pid
+        else 
+            echo "$1 not exit"
+        fi
     fi
 }
 
@@ -24,4 +29,4 @@ function stop(){
     fi
 }
 
-$1
+$1 $2
